added an R script using socialmixr for contact mat
socialmixr_test.R added. Produces a re-scaled, per-capita contact matrix for the age groups [0 - 70) and [70, 70+] from contact data pulled from the POLYMOD survey of the UK. As far as I can tell this can be used directly in the model, as long as beta (the global transmission scalar) is adjusted to ensure the right R0.