added derivation of risk factors for age groups 1 and 2
Calculations are in the csv file: Australia_National_Age_and_Sex_2023.csv
Assumption: the rate 2/1000 is the total rate of adverse events for vaccination. The ratio 10 is the risk in men under 30 relative to all others. This gives a risk of about 9/1000 for men under 30 and 9/10,000 for everyone else. Stratifying by age, for the population under 70 this produces an overall risk of 2.67/1000 and, for the population over 70 it's 9/1000 because it comprises only the low-risk group.