added global loss surfaces to glam plots AND implemented global loss function
implemented heatmaps with seaborne (this also uses matrix input) and added plotting scripts for global loss function vs v1 and v2. IMPORTANT: the global loss function is now explicitly defined in model.py as global_loss(). This should be used instead of any in-line definitions of the loss function (I noticed this crucial formula was repeated in several places, so please use the function rather than the in-line implementation).