attempting to plot global loss function
included global loss function, trying to make sure normalisation is accurate (i.e., representative of what actually happens in the optimisation routine). This needs to be addressed in optimisation.py and model.py where the loss terms and normalisation are performed - currently the normalisation transform is implemented in two separate places.