ran simulations for unlimited vaccination, R0 = 1.5
Generated the database and figures, added a new config file.

WARNING: turned off the error message about simulation timing in the database generator script (I made a new file with the modified code), I ran into a memory issue, basically couldn't run the simulations long enough, got above 20,000 days and ran out of memory. Results look fine, but we should check that the herd immunity thresholds are where we expect them to be. The overall herd immunity threshold seems a bit low, given that efficacy against transmission is only 50%ish, but this could be a result of the population contact structure.