added some additional heatmaps - tested for R0 = 1.5
added a script to plot some additional heatmaps of burden components to expand on the results we got for low R0. I've also had a stab at computing the final size for R0=1 and plotting this to represent the herd immunity threshold on the heatmap of total infections. The calculations are in the matlab script final_size.m which is in the extrascripts directory. Someone should check the calculations to make sure the equations are correct (the result seems reasonable but it would be good to get another set of eyes on it).