added some scripts for plotting trajectoris, removed old configs
added a script for plotting infection trajectories for specified values of v1 and v2, also modified the plotting script for burden surfaces to remove some of the redundant or less useful information (focusing on infection numbers in this version). NOTE: there may be an issue with how beta is computed from R0 and the contact matrix, some of the vaccination results don't make sense, especially when N2 is set to be very small (see out/2025-03-21_test_unlimited_low_R0/glamorous-burden_surfaces.png for an example of this behaviour - the corresponding config file sets N1 = 4999000 and N2 = 1000, but infection rates still depend very stongly on p2).