major change (fix?) to create-grid-database.py and model.py
NOTICE. I think the NGM formulation and the ODEs were both incorrect, and I have tried to fix the issues. This commit includes the population sizes N1 and N2 in the equation for beta (the NGM derivation) and REMOVES them from the ODEs in model.py (lines 321 - 343). I've left the original code commented out for comparison. I tested this new formulation both for the low-R0 case (where it gave sensible results - the herd immunity threshold in v1 no longer depends sensitively on v2, also the critical threshold is in the right spot for our vaccine efficacy assumption). I also tested it on the R0 we used in the manuscript, 3.4, and it changed the results quantitatively to some extent, but not qualitatively (at least not for the loss surfaces, we'll have to check the heatmaps over w_EI and w_EV though).