added an additional manuscript figure (no source code)
added a figure with heatmaps of clinical burden for optimal vacc strategies vs w_EI and w_EV for the three R0 scenarios. This will be useful in the end of the results section. This is a custom figure, built from components drawn from the various output files - there is no associated source code.